PWA: Add alternative prism app icon
