+++ assets/templates/manifest.json
@@ -4,77 +4,77 @@
   "description": "{{ .config.SiteDescription }}",
   "icons": [
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-16.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-16.png",
       "sizes": "16x16",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-32.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-32.png",
       "sizes": "32x32",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-72.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-72.png",
       "sizes": "72x72",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-114.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-114.png",
       "sizes": "114x114",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-128.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-128.png",
       "sizes": "128x128",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-144.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-144.png",
       "sizes": "144x144",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-152.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-152.png",
       "sizes": "152x152",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-160.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-160.png",
       "sizes": "160x160",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-167.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-167.png",
       "sizes": "167x167",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-180.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-180.png",
       "sizes": "180x180",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-192.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-192.png",
       "sizes": "192x192",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-196.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-196.png",
       "sizes": "196x196",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-256.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-256.png",
       "sizes": "256x256",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-400.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-400.png",
       "sizes": "400x400",
       "type": "image/png"
     },
     {
-      "src": "{{ .config.StaticUri }}/img/{{ .config.AppIcon }}-512.png",
+      "src": "{{ .config.StaticUri }}/img/icons/{{ .config.AppIcon }}-512.png",
       "sizes": "512x512",
       "type": "image/png"
     }
@@ -82,8 +82,8 @@
   "scope": "{{ .config.BaseUri }}/",
   "start_url": "{{ .config.BaseUri }}/",
   "display": "{{ .config.AppMode }}",
-  "theme_color": "#0d0d0d",
-  "background_color": "#0d0d0d",
+  "theme_color": "#000000",
+  "background_color": "#000000",
   "permissions": [
     "geolocation",
     "downloads",
